move format_tag to testing.helpers
